EVG-14087 clean up duplicate code
