+++ rest/data/commit_queue_test.go
@@ -8,7 +8,6 @@ import (
 	"github.com/evergreen-ci/evergreen/db"
 	"github.com/evergreen-ci/evergreen/model"
 	"github.com/evergreen-ci/evergreen/model/commitqueue"
-	"github.com/evergreen-ci/evergreen/model/distro"
 	"github.com/evergreen-ci/evergreen/model/patch"
 	"github.com/evergreen-ci/evergreen/model/user"
 	restModel "github.com/evergreen-ci/evergreen/rest/model"
@@ -350,37 +349,6 @@ func (s *CommitQueueSuite) TestMockCommitQueueClearAll() {
 	clearedCount, err := s.ctx.CommitQueueClearAll()
 	s.NoError(err)
 	s.Equal(2, clearedCount)
-}
-
-func (s *CommitQueueSuite) TestAddMergeTaskAndVariant() {
-	config, err := evergreen.GetConfig()
-	config.CommitQueue.MergeTaskDistro = "d"
-	s.NoError(config.CommitQueue.Set())
-	s.NoError(err)
-	s.NoError(db.ClearCollections(distro.Collection))
-	s.NoError((&distro.Distro{
-		Id: config.CommitQueue.MergeTaskDistro,
-	}).Insert())
-
-	project := &model.Project{}
-	patchDoc := &patch.Patch{}
-	ref := &model.ProjectRef{}
-
-	s.NoError(addMergeTaskAndVariant(patchDoc, project, ref, commitqueue.SourceDiff))
-
-	s.Require().Len(patchDoc.BuildVariants, 1)
-	s.Equal(evergreen.MergeTaskVariant, patchDoc.BuildVariants[0])
-	s.Require().Len(patchDoc.Tasks, 1)
-	s.Equal(evergreen.MergeTaskName, patchDoc.Tasks[0])
-
-	s.Require().Len(project.BuildVariants, 1)
-	s.Equal(evergreen.MergeTaskVariant, project.BuildVariants[0].Name)
-	s.Require().Len(project.BuildVariants[0].Tasks, 1)
-	s.True(project.BuildVariants[0].Tasks[0].CommitQueueMerge)
-	s.Require().Len(project.Tasks, 1)
-	s.Equal(evergreen.MergeTaskName, project.Tasks[0].Name)
-	s.Require().Len(project.TaskGroups, 1)
-	s.Equal(evergreen.MergeTaskGroup, project.TaskGroups[0].Name)
 }
 
 func (s *CommitQueueSuite) TestWritePatchInfo() {
