+++ rest/model/schema/type_mapping.yml
@@ -5,3 +5,5 @@ models:
     model: github.com/evergreen-ci/evergreen/model/patch.DisplayTask
   BaseTaskInfo:
     model: github.com/evergreen-ci/evergreen/model/task.BaseTaskInfo
+  Module:
+    model: github.com/evergreen-ci/evergreen/model/commitqueue.Module
