+++ units/commit_queue.go
@@ -4,11 +4,9 @@ import (
 	"context"
 	"fmt"
 	"strconv"
-	"strings"
 	"time"
 
 	"github.com/evergreen-ci/evergreen"
-	"github.com/evergreen-ci/evergreen/db"
 	"github.com/evergreen-ci/evergreen/model"
 	"github.com/evergreen-ci/evergreen/model/commitqueue"
 	"github.com/evergreen-ci/evergreen/model/event"
@@ -321,9 +319,7 @@ func (j *commitQueueJob) TryUnstick(ctx context.Context, cq *commitqueue.CommitQ
 	return
 }
 
-// TODO EVG-14087: delete in favor of copy in rest/data
 func (j *commitQueueJob) processGitHubPRItem(ctx context.Context, cq *commitqueue.CommitQueue, nextItem commitqueue.CommitQueueItem, projectRef *model.ProjectRef, githubToken string) {
-	var patchDoc *patch.Patch
 	pr, dequeue, err := checkPR(ctx, githubToken, nextItem.Issue, projectRef.Owner, projectRef.Repo)
 	if err != nil {
 		j.logError(err, "PR not valid for merge", nextItem)
@@ -336,98 +332,24 @@ func (j *commitQueueJob) processGitHubPRItem(ctx context.Context, cq *commitqueu
 		return
 	}
 
-	if nextItem.PatchId == "" {
-		patchDoc, err = patch.MakeNewMergePatch(pr, projectRef.Id, evergreen.CommitQueueAlias, pr.GetTitle(), nextItem.MessageOverride)
-		if err != nil {
-			j.logError(err, "can't make patch", nextItem)
-			j.AddError(sendCommitQueueGithubStatus(j.env, pr, message.GithubStateFailure, "can't make patch", ""))
-			j.dequeue(cq, nextItem)
-			return
-		}
-
-		patch, patchSummaries, projectConfig, err := getPatchInfo(ctx, githubToken, patchDoc)
-		if err != nil {
-			j.logError(err, "can't get patch info", nextItem)
-			j.AddError(sendCommitQueueGithubStatus(j.env, pr, message.GithubStateFailure, "can't make patch", ""))
-			j.dequeue(cq, nextItem)
-			return
-		}
-
-		errs := validator.CheckProjectSyntax(projectConfig)
-		errs = append(errs, validator.CheckProjectSettings(projectConfig, projectRef)...)
-		catcher := grip.NewBasicCatcher()
-		for _, validationErr := range errs.AtLevel(validator.Error) {
-			catcher.Add(validationErr)
-		}
-		if catcher.HasErrors() {
-			update := NewGithubStatusUpdateJobForProcessingError(
-				commitqueue.GithubContext,
-				pr.Base.User.GetLogin(),
-				pr.Base.Repo.GetName(),
-				pr.Head.GetRef(),
-				InvalidConfig,
-			)
-			update.Run(ctx)
-			j.AddError(update.Error())
-			j.logError(catcher.Resolve(), "invalid config file", nextItem)
-			j.AddError(sendCommitQueueGithubStatus(j.env, pr, message.GithubStateFailure, "can't make patch", ""))
-			j.dequeue(cq, nextItem)
-			return
-		}
-
-		if err = writePatchInfo(patchDoc, patchSummaries, patch); err != nil {
-			j.logError(err, "can't make patch", nextItem)
-			j.AddError(sendCommitQueueGithubStatus(j.env, pr, message.GithubStateFailure, "can't make patch", ""))
-			j.dequeue(cq, nextItem)
-			return
-		}
-
-		modulePRs, modulePatches, dequeue, err := getModules(ctx, githubToken, nextItem, projectConfig)
-		if err != nil {
-			j.logError(err, "can't get modules", nextItem)
-			if dequeue {
-				j.AddError(sendCommitQueueGithubStatus(j.env, pr, message.GithubStateFailure, "can't get modules", ""))
-				j.dequeue(cq, nextItem)
-			}
-			return
-		}
-		patchDoc.Patches = append(patchDoc.Patches, modulePatches...)
-
-		// populate tasks/variants matching the commitqueue alias
-		projectConfig.BuildProjectTVPairs(patchDoc, patchDoc.Alias)
-
-		if err = addMergeTaskAndVariant(patchDoc, projectConfig, projectRef, commitqueue.SourcePullRequest); err != nil {
-			j.logError(err, "can't set patch project config", nextItem)
-			event.LogCommitQueueEnqueueFailed(nextItem.Issue, err)
-			j.dequeue(cq, nextItem)
-			return
-		}
-
-		if err = patchDoc.Insert(); err != nil {
-			j.logError(err, "can't insert patch", nextItem)
-			j.dequeue(cq, nextItem)
-			j.AddError(sendCommitQueueGithubStatus(j.env, pr, message.GithubStateFailure, "can't make patch", ""))
-			return
-		}
-
-		// TODO EVG-14087 also do this when we finalize, with a different message
-		for _, modulePR := range modulePRs {
-			j.AddError(sendCommitQueueGithubStatus(j.env, modulePR, message.GithubStatePending, "added to queue", patchDoc.Id.Hex()))
-		}
-	} else {
-		var err error
-		patchDoc, err = patch.FindOneId(nextItem.PatchId)
-		if err != nil {
-			j.AddError(errors.Wrapf(err, "error finding patch '%s'", nextItem.Version))
-			return
-		}
-		if patchDoc == nil {
-			j.AddError(errors.Errorf("patch '%s' not found", nextItem.Version))
-			return
-		}
+	patchDoc, err := patch.FindOneId(nextItem.PatchId)
+	if err != nil {
+		j.AddError(errors.Wrapf(err, "error finding patch '%s'", nextItem.Version))
+		j.dequeue(cq, nextItem)
+		return
+	}
+	if patchDoc == nil {
+		j.AddError(errors.Errorf("patch '%s' not found", nextItem.Version))
+		j.dequeue(cq, nextItem)
+		return
+	}
+	projectConfig, _, err := model.GetPatchedProject(ctx, patchDoc, githubToken)
+	if err != nil {
+		j.logError(err, "problem getting patched project", nextItem)
+		j.dequeue(cq, nextItem)
+		j.AddError(sendCommitQueueGithubStatus(j.env, pr, message.GithubStateFailure, "can't get project config", ""))
 	}
 
-	// TODO EVG-14087 delete above
 	v, err := model.FinalizePatch(ctx, patchDoc, evergreen.MergeTestRequester, githubToken)
 	if err != nil {
 		j.logError(err, "can't finalize patch", nextItem)
@@ -444,6 +366,16 @@ func (j *commitQueueJob) processGitHubPRItem(ctx context.Context, cq *commitqueu
 	}
 
 	j.AddError(sendCommitQueueGithubStatus(j.env, pr, message.GithubStatePending, "preparing to test merge", v.Id))
+	modulePRs, _, err := model.GetModulesFromPR(ctx, githubToken, patchDoc.GithubPatchData.PRNumber, nextItem.Modules, projectConfig)
+	if err != nil {
+		j.logError(err, "can't get modules", nextItem)
+		j.AddError(sendCommitQueueGithubStatus(j.env, pr, message.GithubStateFailure, "can't get modules", ""))
+		j.dequeue(cq, nextItem)
+		return
+	}
+	for _, modulePR := range modulePRs {
+		j.AddError(sendCommitQueueGithubStatus(j.env, modulePR, message.GithubStatePending, "preparing to test merge", patchDoc.Id.Hex()))
+	}
 
 	event.LogCommitQueueStartTestEvent(v.Id)
 }
@@ -471,7 +403,7 @@ func (j *commitQueueJob) processCLIPatchItem(ctx context.Context, cq *commitqueu
 		return
 	}
 
-	if err = addMergeTaskAndVariant(patchDoc, project, projectRef, commitqueue.SourceDiff); err != nil {
+	if err = AddMergeTaskAndVariant(patchDoc, project, projectRef, commitqueue.SourceDiff); err != nil {
 		j.logError(err, "can't set patch project config", nextItem)
 		event.LogCommitQueueEnqueueFailed(nextItem.Issue, err)
 		j.dequeue(cq, nextItem)
@@ -542,72 +474,6 @@ func checkPR(ctx context.Context, githubToken, issue, owner, repo string) (*gith
 	return pr, false, nil
 }
 
-func getModules(ctx context.Context, githubToken string, nextItem commitqueue.CommitQueueItem, projectConfig *model.Project) ([]*github.PullRequest, []patch.ModulePatch, bool, error) {
-	var modulePRs []*github.PullRequest
-	var modulePatches []patch.ModulePatch
-	for _, mod := range nextItem.Modules {
-		module, err := projectConfig.GetModuleByName(mod.Module)
-		if err != nil {
-			return nil, nil, true, errors.Wrapf(err, "can't get module for module name '%s'", mod.Module)
-		}
-		owner, repo, err := thirdparty.ParseGitUrl(module.Repo)
-		if err != nil {
-			return nil, nil, true, errors.Wrapf(err, "module '%s' misconfigured (malformed URL)", mod.Module)
-		}
-
-		pr, dequeue, err := checkPR(ctx, githubToken, mod.Issue, owner, repo)
-		if err != nil {
-			return nil, nil, dequeue, errors.Wrap(err, "PR not valid for merge")
-		}
-		modulePRs = append(modulePRs, pr)
-		githash := pr.GetMergeCommitSHA()
-
-		modulePatches = append(modulePatches, patch.ModulePatch{
-			ModuleName: mod.Module,
-			Githash:    githash,
-			PatchSet: patch.PatchSet{
-				Patch: mod.Issue,
-			},
-		})
-	}
-
-	return modulePRs, modulePatches, false, nil
-}
-
-func getPatchInfo(ctx context.Context, githubToken string, patchDoc *patch.Patch) (string, []thirdparty.Summary, *model.Project, error) {
-	patchContent, summaries, err := thirdparty.GetGithubPullRequestDiff(ctx, githubToken, patchDoc.GithubPatchData)
-	if err != nil {
-		return "", nil, nil, errors.Wrap(err, "can't get diff")
-	}
-
-	// fetch the latest config file
-	config, projectYaml, err := model.GetPatchedProject(ctx, patchDoc, githubToken)
-	if err != nil {
-		return "", nil, nil, errors.Wrap(err, "can't get remote config file")
-	}
-
-	patchDoc.PatchedConfig = projectYaml
-	return patchContent, summaries, config, nil
-}
-
-func writePatchInfo(patchDoc *patch.Patch, patchSummaries []thirdparty.Summary, patchContent string) error {
-	patchFileID := fmt.Sprintf("%s_%s", patchDoc.Id.Hex(), patchDoc.Githash)
-	if err := db.WriteGridFile(patch.GridFSPrefix, patchFileID, strings.NewReader(patchContent)); err != nil {
-		return errors.Wrap(err, "failed to write patch file to db")
-	}
-
-	// no name for the main patch
-	patchDoc.Patches = append(patchDoc.Patches, patch.ModulePatch{
-		Githash: patchDoc.Githash,
-		PatchSet: patch.PatchSet{
-			PatchFileId: patchFileID,
-			Summary:     patchSummaries,
-		},
-	})
-
-	return nil
-}
-
 func sendCommitQueueGithubStatus(env evergreen.Environment, pr *github.PullRequest, state message.GithubState, description, versionID string) error {
 	sender, err := env.GetSender(evergreen.SenderGithubStatus)
 	if err != nil {
@@ -652,7 +518,7 @@ func validateBranch(branch *github.Branch) error {
 	return nil
 }
 
-func addMergeTaskAndVariant(patchDoc *patch.Patch, project *model.Project, projectRef *model.ProjectRef, source string) error {
+func AddMergeTaskAndVariant(patchDoc *patch.Patch, project *model.Project, projectRef *model.ProjectRef, source string) error {
 	settings, err := evergreen.GetConfig()
 	if err != nil {
 		return errors.Wrap(err, "error retrieving Evergreen config")
